# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  domain.CreateOrderRequest:
    properties:
      shipping_address:
        $ref: '#/definitions/domain.ShippingAddress'
    required:
    - shipping_address
    type: object
...
  domain.OrderStatus:
    enum:
    - pending
    - completed
    - cancelled
    type: string
    x-enum-varnames:
    - OrderStatusPending
    - OrderStatusCompleted
    - OrderStatusCancelled
...
  domain.ShippingAddress:
    properties:
      city:
        type: string
      state:
        type: string
      street:
        type: string
      zip_code:
        type: string
    required:
    - city
    - state
    - street
    - zip_code
    type: object
...
  domain.UpdateOrderStatusRequest:
    properties:
      status:
        allOf:
        - $ref: '#/definitions/domain.OrderStatus'
        enum:
        - pending
        - completed
        - cancelled
    required:
    - status
    type: object
  ecommerce-backend_internal_delivery_http.Response:
    properties:
      data: {}
      error:
        $ref: '#/definitions/http.Error'
      message:
        type: string
      meta:
        $ref: '#/definitions/http.Meta'
      success:
        type: boolean
    type: object
  http.Error:
    properties:
      code:
        type: string
      details: {}
    type: object
  http.Meta:
    properties:
      page:
        type: integer
      per_page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
host: localhost:8080
info:
  contact: {}
  description: This is the backend API for E-Commerce platform
  title: E-Commerce API
  version: "1.0"
paths:
  /admin/orders/{id}/status:
    put:
      consumes:
      - application/json
      description: Update the status of an order (Admin only)
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      - description: Status details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.UpdateOrderStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ecommerce-backend_internal_delivery_http.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ecommerce-backend_internal_delivery_http.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ecommerce-backend_internal_delivery_http.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ecommerce-backend_internal_delivery_http.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ecommerce-backend_internal_delivery_http.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ecommerce-backend_internal_delivery_http.Response'
      security:
      - BearerAuth: []
      summary: Update order status
      tags:
      - admin
...
  /orders:
    get:
      consumes:
      - application/json
      description: Get a paginated list of orders for the authenticated user
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page
        in: query
        name: per_page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ecommerce-backend_internal_delivery_http.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ecommerce-backend_internal_delivery_http.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ecommerce-backend_internal_delivery_http.Response'
      security:
      - BearerAuth: []
      summary: List user orders
      tags:
      - orders
    post:
      consumes:
      - application/json
      description: Creates an order from the current user's shopping cart. Will fail
        if cart is empty or stock is insufficient.
      parameters:
      - description: Shipping Address details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.CreateOrderRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/ecommerce-backend_internal_delivery_http.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ecommerce-backend_internal_delivery_http.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ecommerce-backend_internal_delivery_http.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ecommerce-backend_internal_delivery_http.Response'
      security:
      - BearerAuth: []
      summary: Create a new order (Checkout)
      tags:
      - orders
  /orders/{id}:
    get:
      consumes:
      - application/json
      description: Get specific order details by ID
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ecommerce-backend_internal_delivery_http.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ecommerce-backend_internal_delivery_http.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ecommerce-backend_internal_delivery_http.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ecommerce-backend_internal_delivery_http.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ecommerce-backend_internal_delivery_http.Response'
      security:
      - BearerAuth: []
      summary: Get order details
      tags:
      - orders
  /orders/{id}/cancel:
    put:
      consumes:
      - application/json
      description: Cancel an order that is still in pending status
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ecommerce-backend_internal_delivery_http.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ecommerce-backend_internal_delivery_http.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ecommerce-backend_internal_delivery_http.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ecommerce-backend_internal_delivery_http.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ecommerce-backend_internal_delivery_http.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ecommerce-backend_internal_delivery_http.Response'
      security:
      - BearerAuth: []
      summary: Cancel an order
      tags:
      - orders
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"